修改__protect_rm函数,增加判断

+++ zsh/alias.zsh
@@ -40,13 +40,21 @@ alias g:r="php artisan generate:resource"
 
 # rm
 function __protect_rm {
-	local sub_dir=`date +%y_%m_%d_%H_%M_%S`
+	local sub_dir=`date +%Y/%m/%d_%H_%M_%S`
 
 	[[ ${1:0:1} == "-" ]] && shift
 
+	[ "$#" = 0 ] && return 1
+
 	local dir=~/.Trash/rmData/$sub_dir
 	[[ -d $dir ]] || mkdir -p $dir
-	echo "rm $@" >> $dir/__command.txt
 	mv --target-directory $dir "$@"
+	if [ $? -eq 0 ]; then
+		echo "rm $@" >> $dir/__command.txt
+	else
+		echo "error code: "$? >> $dir/__error.txt
+		return 2
+	fi
+	return 0
 }
 alias rm='__protect_rm'
